fix: PostgreSQL numeric 오류로 cost 타입을 DOUBLE PRECISION으로 수정 (fix PostgreSQL schema error by changing cost from DECIMAL to DOUBLE PRECISION)

+++ backend/src/main/resources/db/migration/V3__refactor_to_execution_call.sql
@@ -19,7 +19,7 @@ CREATE TABLE executions (
 
     -- Aggregated metrics from calls
     calls_count INTEGER DEFAULT 0,
-    total_cost DECIMAL(10, 6),
+    total_cost DOUBLE PRECISION,
     total_tokens BIGINT
 );
 
@@ -41,7 +41,7 @@ CREATE TABLE calls (
     completion_tokens INTEGER,
     total_tokens INTEGER,
 
-    cost DECIMAL(10, 6),
+    cost DOUBLE PRECISION,
     latency_ms BIGINT NOT NULL,
 
     status VARCHAR(20) NOT NULL,
